+++ src/main/java/backend/wal/user/domain/aggregate/entity/User.java
@@ -3,6 +3,9 @@ package backend.wal.user.domain.aggregate.entity;
 import backend.wal.onboarding.domain.entity.Onboarding;
 import backend.wal.user.domain.aggregate.vo.SocialInfo;
 import backend.wal.user.app.dto.request.CreateUserDto;
+import backend.wal.user.domain.aggregate.vo.SocialType;
+import backend.wal.user.domain.aggregate.vo.UserRole;
+import backend.wal.user.domain.aggregate.vo.UserStatus;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
