+++ src/main/java/backend/wal/user/domain/aggregate/vo/SocialInfo.java
@@ -1,6 +1,5 @@
 package backend.wal.user.domain.aggregate.vo;
 
-import backend.wal.user.domain.aggregate.entity.SocialType;
 import com.google.common.base.Objects;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
