+++ src/main/java/backend/wal/user/domain/aggregate/vo/SocialType.java
@@ -1,13 +1,15 @@
-package backend.wal.user.domain.aggregate.entity;
+package backend.wal.user.domain.aggregate.vo;
 
 import lombok.AccessLevel;
+import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+@Getter
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
-public enum UserRole {
+public enum SocialType {
 
-    ADMIN("관리자"),
-    USER("일반 유저"),
+    APPLE("애플"),
+    KAKAO("카카오"),
     ;
 
     private final String value;
