+++ src/main/java/backend/wal/user/domain/aggregate/vo/UserRole.java
@@ -1,15 +1,13 @@
-package backend.wal.user.domain.aggregate.entity;
+package backend.wal.user.domain.aggregate.vo;
 
 import lombok.AccessLevel;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
-@Getter
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
-public enum SocialType {
+public enum UserRole {
 
-    APPLE("애플"),
-    KAKAO("카카오"),
+    ADMIN("관리자"),
+    USER("일반 유저"),
     ;
 
     private final String value;
