+++ src/main/java/backend/wal/user/domain/aggregate/vo/UserStatus.java
@@ -1,4 +1,4 @@
-package backend.wal.user.domain.aggregate.entity;
+package backend.wal.user.domain.aggregate.vo;
 
 public enum UserStatus {
     ACTIVE,
