+++ src/test/java/backend/wal/auth/app/service/provider/AuthServiceProviderTest.java
@@ -3,7 +3,7 @@ package backend.wal.auth.app.service.provider;
 import backend.wal.auth.app.service.AppleAuthService;
 import backend.wal.auth.app.service.AuthService;
 import backend.wal.auth.app.service.KakaoAuthService;
-import backend.wal.user.domain.aggregate.entity.SocialType;
+import backend.wal.user.domain.aggregate.vo.SocialType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
